refactor: split jira-client.ts into smaller, focused modules
- Created new directory structure under src/lib/jira-client/
- Extracted schemas and types into jira-client-types.ts (96 lines)
- Extracted base client class into jira-client-base.ts (25 lines)
- Extracted issue operations into jira-client-issues.ts (600 lines)
- Extracted comment operations into jira-client-comments.ts (188 lines)
- Extracted board operations into jira-client-boards.ts (378 lines)
- Extracted sprint operations into jira-client-sprints.ts (243 lines)
- Extracted user operations into jira-client-users.ts (75 lines)
- Main jira-client.ts now serves as a facade that composes all modules (187 lines)

This refactoring improves code organization and maintainability by:
- Keeping each file under 700 lines as required
- Grouping related functionality together
- Maintaining the existing public API unchanged
- Preserving all TypeScript types and Effect-based implementations

All functionality has been preserved and tested - the CLI continues to work correctly.
