+++ src/lib/jira-client/index.ts
@@ -1,0 +1,7 @@
+export * from './jira-client-types.js';
+export * from './jira-client-base.js';
+export * from './jira-client-issues.js';
+export * from './jira-client-comments.js';
+export * from './jira-client-boards.js';
+export * from './jira-client-sprints.js';
+export * from './jira-client-users.js';
